Crate pallet_executive

Executive Module

This is a fork of the upstream frame-executive. Most of the superior features have been removed, this pallet aims to provide the minimum stable interfaces for runtime.

The Executive module acts as the orchestration layer for the runtime. It dispatches incoming extrinsic calls to the respective modules in the runtime.

Overview

The executive module is not a typical pallet providing functionality around a specific feature. It is a cross-cutting framework component for the FRAME. It works in conjunction with the FRAME System module to perform these cross-cutting functions.

The Executive module provides functions to:

Check transaction validity.

Initialize a block.

Apply extrinsics.

Execute a block.

Finalize a block.

Start an off-chain worker.

The flow of their application in a block is explained in the block flowchart.

Implementations

The Executive module provides the following implementations:

ExecuteBlock: Trait that can be used to execute a block.

Executive: Type that can be used to make the FRAME available from the runtime.

Modules

block_flowchart

Block Execution

Structs

Executive

Main entry point for certain runtime actions as e.g. execute_block.

Type Aliases

CallOf

CheckedOf

OriginOf